@@ -10,10 +10,10 @@
     ],
     "minimum-stability": "stable",
     "require": {
-        "symfony/event-dispatcher": "~3"
+        "symfony/event-dispatcher": "~4"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4"
+        "phpunit/phpunit": "^6"
     },
     "autoload": {
         "psr-0": {
